Update IdGenerator
summary:
- pid-long conversion methods were moved to class IdGenerator
- added an overloaded sequentialMint(long amount, long startingValue)
- added calculatePermutations has been renamed to getMaxPermutation
- remove unused imports
